added getReviews API

+++ backend/database/migrations/2024_03_31_181241_create_ride_reviews_table.php
@@ -13,7 +13,7 @@ return new class extends Migration
     {
         Schema::create('reviews', function (Blueprint $table) {
             $table->id();
-            $table->integer("rating");
+            $table->float("rating");
             $table->text("text");
             $table->unsignedBigInteger("user_id");
             $table->foreign("user_id")->references("id")->on("users")->cascadeOnDelete()->cascadeOnUpdate();
